+++ src/admin/database/migrations/2025_11_09_180000_add_points_to_customers_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->integer('points')->default(0)->after('balance');
+            $table->decimal('points', 15, 4)->default(0)->after('balance');
         });
     }
